CHEF-73 refactor: STT 모듈 구조 개선
- User Session 모듈 리팩토링 (src/user_session/)
  - UserSession으로 유저와 세션 관련한 기능 이동

+++ src/user_session/service.py
@@ -1,7 +1,7 @@
 import uuid
 from fastapi import WebSocket
-from uvicorn.main import logger
-from .exceptions import SessionErrorCode, SessionException
+
+from src.intent.models import Intent
 from .repository import UserSessionRepository
 
 from .client import UserSessionClient
@@ -12,23 +12,21 @@ class UserSessionService:
         self.repository = repository
         self.client = client
 
-    async def create(self, client_websocket: WebSocket):
+    async def create(self, client_websocket: WebSocket) -> str:
         session_id = uuid.uuid4().hex
-        try:
-            self.repository.create_session(session_id, client_websocket)
-            return session_id
-        except SessionException as e:
-            logger.error(f"[ClientSessionService] {e.code.value}: {e.original_exception}", exc_info=True)
-            raise SessionException(SessionErrorCode.SESSION_SERVICE_ERROR, e)
-        except Exception as e:
-            logger.error(f"[ClientSessionService]: {e}", exc_info=True)
-            raise SessionException(SessionErrorCode.SESSION_SERVICE_ERROR, e)
-            
+        self.repository.create_session(session_id, client_websocket)
+        return session_id
+
     async def remove(self, session_id: str):
-        client_websocket = self.repository.get_user_session(session_id)
-        await self.client.close(client_websocket)
-        self.repository.remove_session(session_id)
+        if self.repository.is_session_exists(session_id):
+            client_websocket = self.repository.get_user_session(session_id)
+            await self.client.close_session(client_websocket)
+            self.repository.remove_session(session_id)
 
     async def send_error(self, session_id: str, error: Exception):
         client_websocket = self.repository.get_user_session(session_id)
         await self.client.send_error(client_websocket, error)
+
+    async def send(self, session_id: str, result: Intent):
+        client_websocket = self.repository.get_user_session(session_id)
+        await self.client.send_result(client_websocket, result)
